[Feat] #80 - MenuCell 구현
